+++ libcontainer/cgroups/systemd/dbus.go
@@ -1,4 +1,4 @@
-// +build linux
+// +build linux,!no_systemd
 
 package systemd
