+++ libcontainer/configs/cgroup_nosystemd_linux.go
@@ -1,0 +1,7 @@
+// +build linux,no_systemd
+
+package configs
+
+// Empty since we're not building w/ systemd
+type CgroupSD struct {
+}
